Temporarily added machine config files for NanoPI Neo2
linux-sunxi has the support upstreamed but hasn't caught the metadata
layer up to the support for H5 boards.  Adding at least MY board I
have to my metadata accordingly.  This will come out when the
linux-sunxi project catches up or I've rattled out this stuff and do
a pull request on github for it in their layer to upstream.