Removed cruft...
This recipe was for the OpenWRT layer stuff...which is going to
get overhauled to be something more OE-ish (The maintainer is
doing some...odd...things...leveraging some rather advanced features
of OE that few actually use- along with it being busted all to
hell because it doesn't OE-ize...it's just trying to do OpenWRT
in OE, which is bass-ackwards, ultimately.

Since it's broken, we just need to get it out of the way.  We
will revisit all of it...just will be starting from the ground
up on it.  (Yeah, I know, yet another metadata layer...but we
don't need a third or more of the stuff OpenWRT attempts to do
as it's done differently or better in OE, amongst other things.