Added Bundler support for Ruby...
Ah, OE.  Let's just blacklist anything where they run into a problem
where they missed an RDEPENDS for a recipe and nobody gives it any love
(In this case, it was as STUPID (YES...) as missing having zlib in
their RDEPENDS clause...and since nobody was relying on it at that
particular point...it got AXED out of the latest stable and master.
FAIL.  EPIC FAIL.)

Since we need it for something I'm working on...and will likely need
it or something like it for work on Diaspora* and other projects,
it's getting resurrected for now in MY metadata layer and I may
push a patch back in...someday...maybe not.