Clean and proper...looks like it builds right now
Not that I have a handle on how to fully drive this bbclass yet, but
it _*seems*_ to be doing it right for the first time- which means
it should work now like it's "supposed" to.  I suspect that we'll
need to crack this apart into a baseline ruby class and a gem or
bundler build bbclass ultimately.  Won't know until I try to
drive bundler for the first time.  I'm still leery of using
that particular gem...  It does things in a very un-OE way.