Removed the current classes.  This was a hot-mess...
Didn't work right, which makes it no surprise that it got blacklisted
and removed in master as of 6-9 months ago out of meta-openembedded.

Got what might be a "better" answer from some other upstream project.
It doesn't work right either, but that's more because OE/Yocto
yanked the carpet out from underneath them since Dizzy/Jethro- but
it's fixable as a bbappend off of the main ruby class plus a minor
rework on the bbclass for gemhelping.  Should be next moves there.