Tuning chages to get initial boot.
Not everything working.  May have to do a two-step and do a fetch/patch
without the device tree patches for Neo2 Neo-Plus2 (Need to re-work
the naming ANYHOW...) and then dig a bit deeper and make the tree
make sense and work.  One of the USB ports doesn't work, EMAC doesn't
work right now.  Know these are mere DT fails, in the large so we
should be able to clean those up in short order.