Add NanoPi-Neo-Plus2 build support to U-Boot.
Not complete.  It, however, does build to MACHINE right and boots up
with proper support for the Broadcom WIFI (AMPAK module) on the board
and one of the USB ports lit right.  It needs some more work either
from the U-Boot or Kernel perspective, but this is now cleanly USABLE.