Moved Neo2/Neo-Plus2 support over to the start of a FriendlyElec kernel
They're upstreaming their own fixes/etc against the schematic instead of
what appears to be somewhat half-ass (yes) stuff from linux-sunxi's
efforts.

This appears to be half-assed because linux-sunxi is sort-of tracking
the needed changes for the boards and features *DO NOT WORK* for the
things they're "upstreaming" to linus.  USB root hubs don't work,
WiFi for the neo-plus2 doesn't work right.  So, like with RPi's,
the vendor/community makes a closely curated kernel that **WORKS**
so we'll use it instead.  May not be done- no full build yet(yet!).

Committing because this is the path forward anyhow.