Additional kernel work.
Trying out 4.16.2, which is supposed to have most of the stuff that is
"functional" for an H5, even if it's not fully turned on right, set
correctly functional in the kernel.  We probably need to adjust the
device tree entries for all of this- because there's a crapload of
things that just don't work (because there's stuff in there that's
for the OrangePI variants and they have an ever so slightly differing
circuit design outside of the SoC.  FriendlyElec's kernel build
semi-works in other odd ways and the USB's still don't work right
even using their setup what consists of their Device Tree (sort-of
I suspect...).  Mildly annoying- but I'll get it working here.

Just want a good stable baseline even if it's not an LTS kernel
here.