Attempted move to 4.16.6 of kernel
Has the bulk of all of the needed support, unpatched, for the Neo2
and Neo-Plus2 boards from FriendlyElec.  Once up, we may need the
device tree for the Neo2 patched a bit as it's on-board WiFi has
some hinky stuff going on- but we can't even validate that we need
it until we get the kernel up.