Re-arranged boot order...
Unfortunately, linux-sunxi has presumed something of a no-no here.  That
because the OrangePI and NanoPI boards basically share hardware/device
profiles down the line, they have everything on the identical ports.

Nope.  eMMC and SD are **SWAPPED** one to the other.  Need to look at
what all I can roll-up on a pull request to them on this one.  Really
need to customize this per-board if I can get OE to build it that way
since the boards ARE different.