Attempt to get 4.14.15 up and running with U-Boot 2018.03
Still doesn't work, but we're going to hold on to this for now.  If I
actually "need" to bring up FriendlyElec's kernel I'll do that, but
now it's just trying to figure out what the hell is going on that casues
the hang on boot (which is still happening- so it's not Kernel
dependent...)

Some more analysis needed to sort this out, but we're committing this
just so we have the line in sand here for this.