Reverting change.
Seems that I was chasing my tail a bit.  Comes from not being able to
properly dedicate time to this when I needed to.  Wasn't generating
a clean rootfs so it wasn't booting right.  Put it back the way it was
so I should be actually close to a real cut. May move it to 4.16 even
if it's not an LTS kernel right now so we can scoop up all the latest
and greatest for this SoC we're working with.