Removing FriendlyElec's version set for kernel.
Don't need it.  Found out what the last problem was.  Device uses
RF-Kill function incorrectly with GPIO set active **LOW** (Sigh...)
when the line is assert HIGH to enable.  Will be removing this
since the Upstream to Linus version semi-works and the 4.16.6 that
is my own mod of the linux-sunxi repository DOES work like it
is supposed to.  Will be modding the on-board to WiFi enabled out
of box.  All we need do is absorb the rest of the needed device
tree mods for the bulk of what needs to be done.