Removed 4.14 kernel support.
4.16 correctly works for USBs and we just need to figure out how to get
that enable line to work right for the SDIO WiFi.  (The way they went
to do this is faintly hinky and while they got it to "work" for Ubuntu
it doesn't work right here...so there's a hack somewhere leveraging the
power-on sequencing and all that probably shouldn't be used.

Probably will do a gpio-hog line in the .dts file to extend things
so that we init hot and by the time the SDIO is detected and we go
live with the kernel module, it'll just simply work.