Put this line back in for now.
It doesn't work right.  Some of this isn't being fully and properly
honored or it'd be coming up right now (The line is LOW at the time
I get to it on a command prompt and bring the GPIO active to the
sysfs end- which should have it HIGH...)

Need to look at this but need to have a temp line in the sand for
it all.