Initial commit for maybe having XR819 support
It's unstable.  It's shite.  But it does provide some semblance
of WiFi on a few of the Allwinner boards (I honestly wish not, but
it IS what it is...)- this cowpiles right now, but it's close.

Makefile probably needs to be normalized against something like
the Realtek out of tree builds so it will work "right".  Allegedly
from the upstream this compiles instead of cowpiles and works for
up to 4.17 in the main 4.X series.

Checking it in so I have something to refer to and work with.