Made the pwrseq "devices" modular...
The MMC core code has both loaded with the prior defconfig and the same
is pretty broken (i.e. it just simply sits and spins, not working)- how
they seem to work is that it selects which is being used at probe time
and uses it once the core knows that the controller and "card" have
settled to the proper voltage asserting and deasserting "resets
accordingly.

This fixes the Broadcom/Cypress WiFi core in the Ampak AP6212 WiFi
BT device- but does not fix the BT functionality reset or move the
Broadcom WiFi core into high-perf mode (We have the out-of-band
interrupt source wired on the boards that this is relevant to on
my layer support set for FriendlyElec's devices (Which I intend
to have more of...)  The fixes here once I validate that the
WiFi DID come up with the one (not included yet) patch for the
NVRAM firmware that needs to be obtained from somewhere valid, will
be in a follow-on patch.