Fixes for building PROPERLY against native sysroot.
The kernel build had a leak out to host that just got plugged
with this, thanks to 96boards' own fixes for the same for THEIR
specific kernel recipies tuned to their board offerings.
(Kudos to them for that fix.  This should've been in the main
kernel bbclass stuff, but isn't...)