Kludged in a fix for the compilation of the custom U-Boot config.
I shouldn't HAVE to do this, but the way linux-sunxi was building this
they ended up having it peeling the result out of the libpython dirs
on the **HOST** which as we all know is *BAD* for OE and crosscompiles.

Need to think on how to cleanly make the script/envvar production
process separated from this build so we can use the stock build and
then use that means to do this the *RIGHT* way.  This will suffice
for now...