Reorg for image build turn back on RTL8812AU support.
Need to see if the compile fixes corrected the build issues with
RTL8812AU or not.  It's building on my laptop, but not on anything
that doesn't have libpython3 going on it and a few other select
fails.  Suspect I might've mostly fixed things at this point.
Need to see.  Also re-organized a few declarations into more clear
and appropriate groupings within the build spec.