Turned on the pins for UART3 for BT support...
Only one thing left in this case...  One needs to enable the GPIO and
drive it high.  Not sure if I can tie it to the powerseq for the WIFI
or not.  If so, it'll largely take care of itself and you'll only
need to do an hciattach to UART3 and be golden for BT.