Add experimental ath10k USB, etc. support
Got toys to play with.  Not going to worry about production uses of this
unless it shows to be stable- but someone's got to USE this stuff to
feed back to Qualcomm and Silex (since they're the guilty parties for
the USB and SDIO versions of these drivers) to let them know what is
and isn't working right.