Removing that which doesn't fully work right.
The gems layer is craptastic.  While it'd be nice to mannage a shared
pool of gems to reduce a bit of footprint, the proper practice for
production systems since Ruby understands this concept is to have
the gem versions and gems all bundled up into the "app" and to have
pathing, etc. set for Ruby with a launcher shell.

So, we grabbed meta-ruby's contents for the one thing we need to do
this for the stuff WE care about packaging, bundler, since it will
bundle apps for us and you can hand-load anything else you want on
a local machine with the services OE's providing.