Adding Ruby Bundler support.
This seems to do what it's supposed to and allows you to package gems as
needed and lets Bundler actually, you know, run.  It was a simple
issue for the class object and the OE crowd was (yes, I'm going here)
too damned lazy to correct an absolute path reference in a link for
the gem binary install link- so they blacklisted it and killed it.

Heh...fixed and now it's **MINE**  X-D   I'll maintain it and I
_*MIGHT*_ eventually upstream these parts back in.