Correction for compile behavior problems.
We are using a very specific build for the upstream kernel
that works *correctly* with recipe specific sysroots.  This
append makes it work with the kernel patches we're using
in that build to grab the "right" things when we do the
out of tree build as if we were building the kernel.