Correct recipe for forensics image.
Bundler just got added properly from the blacklisted and
removed meta-ruby piece from meta-openembedded where they
didn't bother to move a fixed path to a relative one for
a symlink for inclusion into /usr/bin.  Since that's been
fixed and the old recipes-rubygems died out of this layer
we're moving to it- which works RIGHT on target.