Removed Toxic IM client...
Build is broken and not really worth sorting the mess out.  Toxic has...
some odd notions...about what should be in a console only chat client
including video (Seriously?  If you don't have GPU/Graphics resources
to do the Qt or GTK+ clients...why would you be CARING about video?)

Since build is broken and not worth my time...for now it's coming out.
We'll add the VPN layer stuff, etc. that goes with Toxcore once we
move to latest in it's recipe.