Found a **NEW** crack for the build to fail upon
Something about the dependencies just not being right under all
conditions.  I'm tired of chasing my damn tail here...  May have
to have a FOUL word or four with Purdie and Company over this.

It's all well and good to be "deterministic" but when there's
no good way to tell the idiot thing that it **NEEDS** the
kernel headers and build config and to stage it somewhere, there
is a damn problem.  Will revisit, HOPEFULLY this weekend.
Need to get a client's build out the door and this is fubaring
me hard.