Stripping all the known to have issues builds OUT.
I need to sort out the CRAP (YES, I'm calling it that...) that
is in the Realtek Makefiles so that it's a **CLEAN** out of
tree module build as well as a DKMS build and see what I can
do to sort out the monitor mode and packet injection such
that it works "right" for some of the later versions of this
software.

This shit of where I can't get a consistent bead on anything
out of Realtek on an OE build is getting **VERY** old- and it's
THEIR fault, really.