The stability fix for meta-rtlwifi pieces...
Seems that the kernel drifted from the version we'd specified
to the minor rev difference used by linux-sunxi for meta-sunxi.

Seems it wants include/config/auto.conf in addition to autoconf.h
in the other directory.  Need to stall the meta-rtlwifi build
stages to the post-compile so that this is where they want it to
be when they actually DO the compile.