A more capable re-work that works more consistently...
It still needs proper fixing, but this may entail re-doing a Yocto
recipe piece to do it right.  So, for now, to let this build "right"
so that we can get our wireless drivers, we're going to nab all of
the build dir in the main kernel build area and pour it right
into the shared artifacts place...which may be "right" anyhow.