Cleaned up the Device Tree patchset
It originally grew out of an evolutionary process instead of an explicit
per-patch process because of needs.  Now, we're needing things to work
against a cleaner/clearer patch for the dts/dtb for the board
that includes cooling controls, speed setting, etc. in the right
places and done the right ways.  So we re-spun the patch.