Initial cut of something resembling Bluetooth SUPPORT.
This turns on at 115k and uploads the firmware at the same speed.
In truth, we need to see what will let us turn it all on at 3Mbits
or 4Mbits and all.  Once there, it should just work swiftly.
Putting this up because we need BT support now as opposed to later,
even if it's suboptimal.