Some hackery for clean BT init...
Seems that BT is dodgy for init on the AP6212 (Doesn't work consistent
as currently specified)- this does NOT work, but I need to preserve
the work up to this point in order to fix it.  There's a few nice
things in this that need a bit more work (Like a first-run block for
eudevd where the first rfkill does nothing, and subsequent ones do
the right thing.  We need to hold it in reset and pull it back out
for things to work right...another fubar with THIS particular core...)
Once sorted out, it'll hopefully be simple, graceful, does what
needs to be done right out of box with rfkill driving it, etc.