Re-worked the recipe to account for no BT support...
Not all boards that are ultimately supported by this metadata
layer use an AMPAK or similar BT device on-board.  (Neo2?)
So, ultimately it needs better handling there for those boards.
This would be the start thereof.  It'll need refactoring once
we go to more than just the Neo-Plus2 support.  Later on that
though.