Moved from a deferred FW load for BT to a direct one on boot...
Since there apparently is a sharp edge with the RFKILL/eudev event
processing, we'll skip trying to be fancy and defer FW load so we're
up quicker on boot and boot 15-ish seconds later.  Seems that there
is lingering resource issues with the eudev callback- it seems that
there might be a lingering issue with some of this that causes
a kernel paging problem with at least the Neo-Plus2 target we're
running with.