Adjusted the preferred RTC for PHA projects...
FriendlyElec opted to not use the RTC on the SoC (Sigh...why?)
so we chose for our projects that "need" this functionality
to use the Maxim DS1307 for use.  The kernel provides the
symlink initially on boot to /dev/rtc to whatever is specified
in our config here- which doesn't show unless it is set as being
in the kernel statically (You can do this as a module but you
do a few extra manual steps in the boot process.  For now, we're
going do it as a "Y" to simplify things so it works right out
of box.