WIP : Extension to Ruby's recipe
Ruby...it's a bit of a hot mess, really.  It doesn't build to target
right.  It can't realistically cross-compile right (It cross-compiles
only it's recipe and messes THAT up).

It is our hope to "fix" this and not like you'd fix an animal.

What the current commit does is that it cleans up the rbconfig.rb
file that has all the hooks to tell the rest of the tools how to build
native code binaries for gems, etc.  With the stock, there's no way
in HELL of your getting bundler, gem, etc. to work right on target
even if you have the target native dev environment on the image.
It has all the wrong paths straight out of your OE build in place- which
breaks **EVERYTHING**.  There's a lot more to do with this bbappend, but
this looks like on first blush that it cleans up the stupids on the
target side and lets us capture the original rbconfig.rb that if you
remove a few burning stupids in it, should let you cross-compile
with ruby.bbclass once it's all fit-n-finished up.

If nothing else, this will let it clean up the stupids out of things
on the target if we can't make the rest happen.  It WORKS for what
I'd envisioned as a clean-up so it goes in regardless (since it doesn't
break anything any worse than the other crap did...) so committing.