Hack our bbappend a bit...
Something I knew was a smidge of an issue bit me in the backside there.
"-gnu" is some appending that gets done with the stuff we're using to
build 64-bit build images for NanoPi Neo2/Neo-Plus2's and it doesn't
play nice with 32-bit builds for things like RasPi 2 or 3's...

Sooo...make it overrode specifically for those two boards for now and
fix it properly shortly.  We need the behavior done there so we can
cleanly handle cross-compiles unlike what Poky/OE-Core currently does
with Ruby (Which is busted all to hell on target for doing anything
pulled in and built there...)