Created a distro entry for meta-pha...
Added pha-linux distro.  This turns on IWD and runit support for
poky relative to the guts IN this layer.  Hopefully we can get this
robust enough to turn on in our local.conf rules and to have our
customers USE this.  We're actually a lot closer than one would
think there.  There's a handful of services that need to be
added in meta-runit so we capture the behavioral rules We have
a handful of overrides because of IWD in meta-pha.

It is hoped that we eventually get these into Yocto/Poky proper
and we provide backfill for at least Sumo and maybe Thud for
this stuff unless these mods get into those versions accordingly.