Add basic hooks to mod for IWD support in this Distro.
We're a distro proper right now, said distro forces IWD and runit
as wifi base management and services/init support.  There's a couple
of "oops..." items in neard and connman that this metadata layer
needs to fix, instead of meta-runit (Which is _only_ responsible
for providing services fixes for the stuff in meta/meta-poky.
There will be another layer called meta-runit-openembedded that
copes with the meta-openembedded services)