Provide a functionality for machines that support arch8a+<foo>
CRC and Crypto are proper add-ons and now there's tune options from
Master that work nicely enough, but meta-tegra's WIP referrs to
this wile being "compatible" with Thud and is technically a hybrid
with Master and Thud.  So, we'll copy it off of Master on meta
to here.  It won't break things, but it'll let meta-tegra (and
possibly others that I can make this work with...) work right
with the Jetpack 4.2 set...