Add some of the needed runit based functions...
We're not tying it in yet (YET!) but we're capturing the needed
bits and bobs.  We need to figure out best path for making this
be in but not being activated unless we're going for a runit
based distro instead.  Some of the other stuff as consumers of
meta-runit are more suitable to not caring and we need to figure out
what all I need to add to make this more generic than they are.