Config moves to warrior support now.
We're building with it and we're walking off from sumo support.
This is a clean break and we'll likely support Thud as a branch
that tracks off of this as well as Warrior when the next version
comes out for Yocto shortly.  We'll likely support 2-3 current
versions as branches like Yocto does with updates going for them
and maybe criticals for 5-6.  Need community help or getting paid
fully as the "day job" for this to be any more than that.