Clean-ups for gocryptfs...
While it worked for warrior...it's not the PROPER way of doing this.

Package sets for "libraries" are really better done as source captures
instead of doing "inherit go" builds.  It sort-of mostly works that
way, but the right way is to just pour the whole of things in like
I'm doing with the go-x-<foo> libs.