Add the likely to be cleanest EXT2/3/4 FUSE support
- Add lwext4, one of the most robust filesystem engine libraries
for EXT4 and below for embedded systems.
- Add a fairly extensive FUSE filesystem using the same.

This support is set up for target/native/nativesdk, but has only
been built/validated for target/native.  nativesdk requires ACL
to be .bbappended for the same and then both it and this be validated
for the mode.  Will get to this (want it...it'd be useful.) eventually.