Move to Zeus plus a bit more.
- Re-added EFL layer (don't know why that wasn't in there.)
- Added tip of ELL git.
- Added tip of IWD git.
- Added tip of Connman git.
- Added tip of tvheadend git.
- Added latest of libhdhomerun to support tvheadend, etc.
- Added ver 5.4.47 of linux-raspberrypi to support ELL/IWD.
- Cleaned up/fixed pha-svcs for recent improvements to
  meta-runit.

There's probably some things to be fixed with meta-efl (incl.
a move to latest on EFL, E, moves to Python 3 where needed
(Dunfell makes this a necessity...), and so forth) but it builds
and cleanly runs on a Pi3 and Pi4 and Connman uses IWD.  In a
manner that looks largely correctly.  Need to put Tethering
through it's paces to see what's still left to do right on
a Pi3/Pi4, but that's a differing commit.  Capturing the
good work and proper function at this time.  Moving on to
being a bit more sophisticated and branching out from this
to get good function available.