Several updates.
- Don't need gpsd and don't have any units with a GPS attached for now.
  (Probably will later...just don't need it in the core regardless)
- Updated Connman to the latest stable release out of git.
- Updated ELL to the latest stable release out of git.
- Updated IWD to the latest stable release out of git.

At this stage we will likely be using IWD full time for many, if not
most of the WiFi capable image releases.  It is VERY robust and cleanly
works with Connman for most of the normal, basic behaviors.  There's
still a few gotchas, but they entail using a SINGLE WiFi Phy in multiple
VIF roles.  (This doesn't seem to work right at all.  It doesn't seem
to honor VIF settings as an administrative thing- and you can't command
IWD with the stuff present to add netdevs to the Phy at this time.
(Doesn't mean it doesn't fully do this- it's just that iwctl and Connman
flatly don't do this right now, so there's some limitations.  Honestly?
it's less efficient and only useful in the most compact configs.  Use
TWO or more Phys when we need that sort of thing like Sta/AP mode or
meshing, etc.)