Side-step tegra targets build problem...
With the metadata set I currently have, it didn't want to build
u-boot-tegra correctly.  What was happening is that version.h
wasn't getting symlinked/copied into tools/ which then broke
the build process.  It's working in other metadata sets correctly
so I'm at a bit of a loss- but the workaround seems to have handled
this solidly enough.

- Force a symlink (since there's no risk here and it works) to
  ${S}/include/version.h to ${S}/tools/version.h to allow the
  code to compile instead of cowpile.
- Make an adjustment in the layer to support external annex
  recipes like meta-runit as this only really belongs to
  meta-tegra.
- Updated the workaround for normal u-boot since python has
  up-revved from Sumo to Zeus.