Temp patch for PyPI tarball packages
This conforms a fix for a Python Package Index package that
has a tarball sources package.  This will only work with
that- there will have to be some expansions to allow WHEEL
files to be poured on from the upstream (Can be done, but
this wasn't what I was needing at the time.  There's hints
in one of the next checkins, but it's not productized for
a .bbclass for this yet...YET...